Refactor MessageType and handle file info in the messaging app
Refactored the MessageType enum by moving it from lib_rrttp to the messenger application, along with the associated network file info structure. This change streamlines the organization of the code, placing the message types and file handling closer to where they are used. Additionally, generic traits were introduced in the application layer to increase flexibility and reusability of the code. Dependencies were updated with two new packages (bincode and infer) to handle serialization and mime type inference. Minor revisions include code adjustments in the Connector class and renaming of methods in commands.rs.