Implement message sending and receiving in ConnectionManager
Extended ConnectionManager in the transport layer of the application to support message sending with the addition of SyncSender. Accompanying changes are made in the UI layer to allow sending and receiving of messages using this new feature. The existing code has also been improved for better readability and simplicity. Closing thread connections upon termination is also incorporated.