Add SonarLint configuration file
A new SonarLint configuration file, sonarlint.xml, is added to the project's .idea directory. This configures SonarLint plugin settings particular to this project and defines file exclusion rules to ignore certain files or directories from analysis by the plugin. In this case, a specific rust type definition file in the demo messanger is being excluded.